completed indexer implementation
